Extract skill discovery into lib/skill.ts
Move SKILL_NAME, response headers, and skill loading/index helpers out of
src/index.ts into a dedicated module. Decouple from Hono's Context so the
helpers take an ASSETS Fetcher and base URL directly, matching the rest of
src/lib.
